% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_dictionary.R
\name{calc_doc_sentidict_tbl}
\alias{calc_doc_sentidict_tbl}
\title{FUNCTION_TITLE}
\usage{
calc_doc_sentidict_tbl(
  .doc_str,
  .tok_valid_regex = "[[:alnum:]]",
  .parse_doc_fn = parse_doc_simple,
  .calc_tok_sentidict_tbl_fn = calc_tok_sentidict_tbl
)
}
\arguments{
\item{.doc_str}{DESCRIPTION.}

\item{.tok_valid_regex}{DESCRIPTION.}

\item{.parse_doc_fn}{DESCRIPTION.}

\item{.calc_tok_sentidict_tbl_fn}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
