% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_dictionary.R
\name{calc_doc_tok_sentidict_tbl}
\alias{calc_doc_tok_sentidict_tbl}
\title{FUNCTION_TITLE}
\usage{
calc_doc_tok_sentidict_tbl(
  .doc_vec,
  .sentidict_tbl,
  .brkiter_locale = "de_DE",
  .by_sentence = FALSE
)
}
\arguments{
\item{.doc_vec}{DESCRIPTION.}

\item{.sentidict_tbl}{DESCRIPTION.}

\item{.brkiter_locale}{DESCRIPTION.}

\item{.by_sentence}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
