% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_transformer.R
\name{calc_germansentiment}
\alias{calc_germansentiment}
\title{FUNCTION_TITLE}
\usage{
calc_germansentiment(
  .doc_str,
  .germansentiment_model = load_germansentiment_model()
)
}
\arguments{
\item{.doc_str}{DESCRIPTION.}

\item{.germansentiment_model}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
