% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_dictionary.R
\name{calc_tok_sentidict_tbl}
\alias{calc_tok_sentidict_tbl}
\title{FUNCTION_TITLE}
\usage{
calc_tok_sentidict_tbl(
  .doc_str,
  .sentidict_tbl = sentidict_sws_tbl,
  .parse_doc_fn = parse_doc_simple
)
}
\arguments{
\item{.doc_str}{DESCRIPTION.}

\item{.sentidict_tbl}{DESCRIPTION.}

\item{.parse_doc_fn}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
