% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derp-package.R
\docType{package}
\name{derp-package}
\alias{derp}
\alias{derp-package}
\title{derp: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
