% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_spacy.R
\name{load_spacy_model}
\alias{load_spacy_model}
\title{FUNCTION_TITLE}
\usage{
load_spacy_model(.model_name = "de_core_news_lg")
}
\arguments{
\item{.model_name}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
