% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{load_support_email_tbl}
\alias{load_support_email_tbl}
\title{FUNCTION_TITLE}
\format{
A tibble with 1700 rows and 8 columns:
\describe{
\item{\code{businessScenario}}{character. DESCRIPTION.}
\item{\code{id}}{character. DESCRIPTION.}
\item{\code{date}}{character. DESCRIPTION.}
\item{\code{category}}{character. DESCRIPTION.}
\item{\code{keyword}}{character. DESCRIPTION.}
\item{\code{chunk_id}}{integer. DESCRIPTION.}
\item{\code{chunk_str}}{character. DESCRIPTION.}
\item{\code{chunk_cat}}{character. DESCRIPTION.}
}
}
\usage{
load_support_email_tbl()
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# support_email_tbl <- load_support_email_tbl()
}
