% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_simple.R
\name{parse_doc_simple}
\alias{parse_doc_simple}
\title{FUNCTION_TITLE}
\usage{
parse_doc_simple(.doc_str, .by_sentence = TRUE, .brkiter_locale = "de_DE")
}
\arguments{
\item{.doc_str}{DESCRIPTION.}

\item{.by_sentence}{DESCRIPTION.}

\item{.brkiter_locale}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
