% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_spacy.R
\name{parse_doc_spacy}
\alias{parse_doc_spacy}
\title{FUNCTION_TITLE}
\usage{
parse_doc_spacy(
  .doc_str,
  .spacy_model = load_spacy_model("de_core_news_lg"),
  .n_process = 1L,
  .batch_size = 50L
)
}
\arguments{
\item{.doc_str}{DESCRIPTION.}

\item{.spacy_model}{DESCRIPTION.}

\item{.n_process}{DESCRIPTION.}

\item{.batch_size}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
