% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vec_slice}
\alias{\%0\%}
\alias{\%||\%}
\alias{\%@\%}
\alias{\%@\%<-}
\alias{\%<-\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:op-get-attr]{\%@\%}}, \code{\link[rlang:op-get-attr]{\%@\%<-}}, \code{\link[rlang:op-null-default]{\%||\%}}}

  \item{vctrs}{\code{\link[vctrs:op-empty-default]{\%0\%}}, \code{\link[vctrs]{vec_slice}}}

  \item{zeallot}{\code{\link[zeallot:operator]{\%<-\%}}}
}}

