% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentidict_sws_tbl}
\alias{sentidict_sws_tbl}
\title{sentidict_sws_tbl: Sentiment dictionary based on "SentiWS"}
\format{
A data frame with 34808 rows and 3 variables:
\describe{
\item{\code{tok_str}}{character. DESCRIPTION.}
\item{\code{tok_pol_lab}}{factor. DESCRIPTION.}
\item{\code{tok_pol_num}}{double. DESCRIPTION.}
}
}
\usage{
sentidict_sws_tbl
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
