% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python.R
\name{setup_vns_condaenv}
\alias{setup_vns_condaenv}
\title{FUNCTION_TITLE}
\usage{
setup_vns_condaenv(
  .conda_envname = "vns_condaenv",
  .install_miniconda = NULL,
  .create_condaenv = NULL
)
}
\arguments{
\item{.conda_envname}{DESCRIPTION.}

\item{.install_miniconda}{DESCRIPTION.}

\item{.create_condaenv}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
