% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{load_amazon_review_tbl}
\alias{load_amazon_review_tbl}
\title{FUNCTION_TITLE}
\format{
A tibble with 210000 rows and 4 columns:
\describe{
\item{\code{doc_id}}{character. DESCRIPTION.}
\item{\code{doc_title}}{character. DESCRIPTION.}
\item{\code{doc_text}}{character. DESCRIPTION.}
\item{\code{doc_label_num}}{integer. DESCRIPTION.}
}
}
\usage{
load_amazon_review_tbl()
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
amazon_review_tbl <- load_amazon_review_tbl()
}
