% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_transformer.R
\name{load_germansentiment_model}
\alias{load_germansentiment_model}
\title{FUNCTION_TITLE}
\usage{
load_germansentiment_model()
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
